requirements.txt
chardet>=2.3.0
git+https://github.com/xBeAsTx/python-entropy@win_compatibility
klein>=16.12.0
patch>=1.16
py2neo>=3.1.2
pycryptodome>=3.4.3
py-gfm>=0.1.3
pyjwt>=1.4.2
python-vote-full>=1.0
pyyaml>=3.11
redlock>=1.2.0
stem>=1.5.3
twisted>=16.6.0
txredisapi>=1.4.4
unidiff
urwid>=1.3.1
